bilateral.m is working now
adjusted class of padded image to be the same as class of input image to
avoid problems (in addpadding.m)
bilateral.m is working now. It's pretty slow compared to convolute with
Gaussian for example.
We have to think about how to implement even sigmas.
I sent a mail to them asking if (in 2c) "bilateral range filter" really
means omitting the domain part and only using the range part.